chore: split components
